requirements.txt
google-cloud-storage==3.7.0
google-cloud-bigquery==3.40.0
apache-airflow[google]
python-dotenv==1.2.1